backend/app/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain
langchain-community
groq
python-dotenv
